Moved dist import functions to distribution.py
